requirements.txt
streamlit
openai>=1.43.0
yt-dlp
ffmpeg-python
python-docx
reportlab
srt
webvtt-py
pydub
python-dotenv
pyannote.audio>=3.1
torch
# Optional: torchaudio (for some backends)
# torchaudio 